Move MutatingPattern to IR module
Move most of the implementing classes.
