+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/pipes/SetOperation.scala
@@ -26,7 +26,7 @@ import org.neo4j.cypher.internal.compiler.v3_2.mutation.makeValueNeoSafe
 import org.neo4j.cypher.internal.compiler.v3_2.planner._
 import org.neo4j.cypher.internal.compiler.v3_2.spi.{Operations, QueryContext}
 import org.neo4j.cypher.internal.frontend.v3_2.{CypherTypeException, InvalidArgumentException, SemanticTable}
-import org.neo4j.cypher.internal.ir.v3_2.IdName
+import org.neo4j.cypher.internal.ir.v3_2._
 import org.neo4j.graphdb.{Node, PropertyContainer, Relationship}
 
 import scala.collection.Map
